requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Reporting
reportlab>=4.0.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0

# Optional: Dashboard
streamlit>=1.25.0

# Optional: Advanced analytics
scikit-learn>=1.3.0
statsmodels>=0.14.0